feat: Add script to print Star Wars characters from a specific movie
The code changes include the addition of a script that prints all characters of a Star Wars movie. The script takes a Movie ID as a positional argument and uses the Star Wars API to fetch the characters. The characters are displayed in the same order as the "characters" list in the `/films/` endpoint. The `request` module is used for making API requests.

This commit aligns with the recent repository commits that focus on adding new features and scripts to various projects.